+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [CSE] Improved validation of complex types and mandatory attributes.
+- [DATABASE] Added *TinyDBBufferedStorage* to greatly improve buffered writes to disk.
+
 
 ### Changed